requirements.txt
lxml==4.7.1
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.11
dj-database-url==0.5.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
kombu==5.2.3
nltk==3.6.7
numpy==1.22.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.2
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
textblob==0.17.1
threadpoolctl==3.0.0
tqdm==4.62.3
tweepy==4.5.0
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5